Adding a bunch of conversion tests, fixing them

+++ testbench/modules/testBehavior/languages/L6/source_gen/BHL6/behavior/A__BehaviorDescriptor.java
@@ -35,10 +35,10 @@ public final class A__BehaviorDescriptor extends BaseBHDescriptor {
   }
 
   private static int foo_id5mnatV0hwr3(@NotNull SNode __thisNode__) {
-    return A__BehaviorDescriptor.bar_id5mnatV0hwrh.invoke(__thisNode__);
+    return ((int) A__BehaviorDescriptor.bar_id5mnatV0hwrh.invoke(__thisNode__));
   }
   private static int bar_id5mnatV0hwrh(@NotNull SNode __thisNode__) {
-    return A__BehaviorDescriptor.bar2_id6LtvT_VTCe3.invoke(__thisNode__);
+    return ((int) A__BehaviorDescriptor.bar2_id6LtvT_VTCe3.invoke(__thisNode__));
   }
   private static int bar2_id6LtvT_VTCe3(@NotNull SNode __thisNode__) {
     return TestResults.DEFAULT_RETURN_VALUE;
