+++ testbench/modules/testBehavior/languages/L6/source_gen/BHL6/behavior/E__BehaviorDescriptor.java
@@ -34,7 +34,7 @@ public final class E__BehaviorDescriptor extends BaseBHDescriptor {
   }
 
   private static int foo_id5mnatV0hyuN(@NotNull SAbstractConcept __thisConcept__) {
-    return C__BehaviorDescriptor.foo_id5mnatV0hxLf.invokeSpecial(__thisConcept__);
+    return ((int) C__BehaviorDescriptor.foo_id5mnatV0hxLf.invokeSpecial(__thisConcept__));
   }
   private static Object virtual_id47lrFSh1$Ca(@NotNull SAbstractConcept __thisConcept__) {
     return TestResults.POLYMORPHIC_CHILD;
